refactor(goal): make the goal-change subscription an Effect stream
goalStore owned a fiber and a runtime: subscribeGoalStateChanges forked a
stream drain on the process runtime and returned an interrupt thunk, two
below-boundary Effect.run* sites the ratchet assigns to the Phase 5 tools
lane. The store now owns neither: goalStateChanges(session) is the whole
surface, a Stream of goal mutations on the session's event plane, and each
host forks it at its own R1 boundary.

- Extension: the fork moves to frontend/events/runFactSubscriptions.ts,
  the module that already hosts this idiom for run facts (goal changes are
  session facts). The settings handler is unchanged except its import.
- Desktop: the fork moves to a new thin edge module,
  desktopGoalSubscription.ts, keeping desktopSettingsIpc Promise-native
  (an effect import there would put its existing recovery catches under
  R7's one-pass rule, which is desktop's own catch-row lane).
- The store keeps no runtime import at all; the stream is filter+map over
  the session's events.

Ratchet: Effect.run* row loses goalStore (2 sites); its debtLanes entry is
dropped. No count grows.
