+++ packages/extension/src/frontend/events/runFactSubscriptions.ts
@@ -4,6 +4,7 @@ import { Effect, Fiber, Stream } from 'effect';
 import type { SessionHandle } from '@agent/runtime';
 import { effectRuntime } from '@platform/processRuntime';
 import { aggregateTarget, type AddOutputFilesPayload } from '@shared/schemas';
+import { goalStateChanges, type GoalStateChange } from '@tools/goal';
 
 /** Read a session's `addOutputFiles` facts from now on. */
 export function subscribeAddOutputFilesRunFact(
@@ -19,6 +20,21 @@ export function subscribeAddOutputFilesRunFact(
           filesByRound: event.filesByRound,
         });
       }),
+    ),
+  );
+  return () => {
+    effectRuntime().runFork(Fiber.interrupt(fiber));
+  };
+}
+
+/** Read a session's goal-state changes from now on. */
+export function subscribeGoalStateChanges(
+  session: Pick<SessionHandle, 'events' | 'now'>,
+  listener: (change: GoalStateChange) => void,
+): () => void {
+  const fiber = effectRuntime().runFork(
+    Stream.runForEach(goalStateChanges(session), (change) =>
+      Effect.sync(() => listener(change)),
     ),
   );
   return () => {
